backend/requirement.txt
fastapi
uvicorn
httpx
python-dotenv
httpx
spacy